Use nodemon to restart the server code
