+++ config/webpack/server.js
@@ -42,10 +42,6 @@ let config = {
         loader: 'json-loader'
       },
       {
-        test: /\.jsx$/,
-        loader: 'babel-loader'
-      },
-      {
         test: /\.tsx?$/,
         loader: 'awesome-typescript-loader',
         exclude: /node_modules/
@@ -95,22 +91,6 @@ let config = {
   }
 };
 
-const copySync = (src, dest, overwrite) => {
-  if (overwrite && fs.existsSync(dest)) {
-    fs.unlinkSync(dest);
-  }
-  const data = fs.readFileSync(src);
-  fs.writeFileSync(dest, data);
-};
-
-const createIfDoesntExist = dest => {
-  if (!fs.existsSync(dest)) {
-    fs.mkdirSync(dest);
-  }
-};
-
-createIfDoesntExist('./build');
-createIfDoesntExist('./build/public');
-copySync('./src/favicon.ico', './build/public/favicon.ico', true);
+require('./copy_files');
 
 module.exports = config;
