+++ package.json
@@ -5,12 +5,13 @@
   "main": "build/server.js",
   "scripts": {
     "clean": "rimraf build",
-    "prebuild": "npm run clean",
     "build": "webpack --config config/webpack/index.js",
-    "build:prod": "cross-env NODE_ENV=production npm run build",
-    "postbuild": "webpack --config config/webpack/server.js",
+    "build:prod": "cross-env NODE_ENV=production npm run clean && npm run build",
+    "build:server": "webpack --config config/webpack/server.js",
+    "build:server:w": "npm run build:server -- --watch",
+    "postbuild": "npm run build:server",
     "prestart": "npm run build",
-    "start": "node build/server.js",
+    "start": "nodemon -V -w build/server.js build/server.js",
     "start:prod": "cross-env NODE_ENV=production npm start",
     "test": "karma start config/test/karma.conf.js",
     "lint": "tslint \"src/**/**.ts*\""
@@ -116,6 +117,7 @@
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^2.0.1",
     "mocha": "^3.2.0",
+    "nodemon": "^1.11.0",
     "phantomjs-prebuilt": "^2.1.14",
     "postcss-assets": "^4.1.0",
     "postcss-cssnext": "^2.9.0",
